requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
et_xmlfile==2.0.0
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
numpy==2.2.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
psycopg==3.2.3
pydantic==2.10.3
pydantic_core==2.27.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.32.1
